requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
brotli-python=1.1.0=py312h9f69965_1
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.7.4=hf0a4a13_0
certifi=2024.7.4=pyhd8ed1ab_0
cffi=1.17.0=py312h80c9ed6_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
debugpy=1.8.5=py312h5c2e7bc_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.15.4=pyhd8ed1ab_0
freetype=2.12.1=hadb7bae_2
fsspec=2024.6.1=pyhff2d567_0
gmp=6.3.0=h7bae524_2
gmpy2=2.1.5=py312hfa9fade_1
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.7=pyhd8ed1ab_0
importlib-metadata=8.3.0=pyha770c72_0
importlib_metadata=8.3.0=hd8ed1ab_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.26.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h81bd7bf_0
krb5=1.21.3=h237132a_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240116.2=cxx17_h00cdb27_1
libblas=3.9.0=23_osxarm64_openblas
libcblas=3.9.0=23_osxarm64_openblas
libcxx=18.1.8=h5a72898_4
libdeflate=1.21=h99b78c6_0
libedit=3.1.20191231=hc8eb9b7_2
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=23_osxarm64_openblas
libopenblas=0.3.27=openmp_h517c56d_1
libpng=1.6.43=h091b4b1_0
libprotobuf=4.25.3=hbfab5d5_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.46.0=hfb93653_0
libtiff=4.6.0=hf8409c0_4
libtorch=2.3.1=cpu_generic_h95df8ed_1
libuv=1.48.0=h93a5062_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.16=hf2054a2_0
libzlib=1.3.1=hfb2fe0b_1
llvm-openmp=18.1.8=hde57baf_1
markupsafe=2.1.5=py312he37b823_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h1cfca0a_2
mpmath=1.3.0=pyhd8ed1ab_0
ncurses=6.5=hb89a1cb_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pyhd8ed1ab_1
nomkl=1.0=h5ca1d4c_0
numpy=2.1.0=py312hb544834_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.1=hfb2fe0b_2
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py312h39b1d8d_0
pip=24.2=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=6.0.0=py312h7e5086c_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.5=h30c5eda_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.12=5_cp312
pytorch=2.3.1=cpu_generic_py312h2aa0b4f_1
pyzmq=26.1.1=py312hfa13136_0
readline=8.2=h92ec313_1
requests=2.32.3=pyhd8ed1ab_0
setuptools=72.2.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.6.1=hf6b88df_1
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.2=pypyh2585a3b_103
tk=8.6.13=h5083fa2_1
torchvision=0.18.1=cpu_py312h5f70645_1
tornado=6.4.1=py312h7e5086c_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.2=pyhd8ed1ab_1
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.44.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
zeromq=4.3.5=hcc0f68c_4
zipp=3.20.0=pyhd8ed1ab_0
zstandard=0.23.0=py312h721a963_0
zstd=1.5.6=hb46c0d2_0
